+++ src/main/java/de/themorpheus/edu/taskservice/controller/TaskTypeController.java
@@ -3,6 +3,7 @@ package de.themorpheus.edu.taskservice.controller;
 import de.themorpheus.edu.taskservice.database.model.TaskTypeModel;
 import de.themorpheus.edu.taskservice.database.repository.TaskTypeRepository;
 import de.themorpheus.edu.taskservice.util.ControllerResult;
+import de.themorpheus.edu.taskservice.util.Error;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -20,8 +21,11 @@ public class TaskTypeController {
 		return ControllerResult.of(this.taskTypeRepository.getTaskTypeByNameKeyIgnoreCase(nameKey));
 	}
 
-	public void deleteTaskType(String nameKey) {
+	public ControllerResult<Object> deleteTaskType(String nameKey) {
+		if (!this.taskTypeRepository.existsByNameKey(nameKey)) return ControllerResult.of(Error.NOT_FOUND, nameKey);
+
 		this.taskTypeRepository.deleteTaskTypeByNameKeyIgnoreCase(nameKey);
+		return ControllerResult.empty();
 	}
 
 	public ControllerResult<List<TaskTypeModel>> getAllTaskTypes() {
